refactor(BDD): Add new kubeconnect package

This commit adds new package called kubeconnect and refactored bootstrap
