+++ .gitignore
@@ -5,4 +5,3 @@ nohup.out
 .DS_Store
 .testfiles/
 openebs-operator-autogen.yaml
-.idea
