+++ integration-tests/kubernetes/kubeclient.go
@@ -1,24 +1,22 @@
-package bootstrap
+package kubernetes
 
 import (
 	"flag"
 	"fmt"
 	"os"
 	"path/filepath"
-	"testing"
 
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-	"k8s.io/client-go/kubernetes"
+	kube "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
 )
 
 var (
 	kubeConfigPath *string
-	kubeConfig     *kubernetes.Clientset
+	clientset      *kube.Clientset
 )
 
-func homeDir() (string, error) {
+// HomeDir return the Home Directory of the environement
+func HomeDir() (string, error) {
 	if h := os.Getenv("HOME"); h != "" { // linux
 		return h, nil
 	} else if h := os.Getenv("USERPROFILE"); h != "" { // windows
@@ -27,9 +25,12 @@ func homeDir() (string, error) {
 	return "", fmt.Errorf("Not able to locate home directory")
 }
 
-var _ = BeforeSuite(func() {
-	home, err := homeDir()
-	Expect(err).To(BeNil())
+// GetClientSet returns the clientset for interacting the kubernetes cluster
+func GetClientSet() (cl *kube.Clientset, err error) {
+	home, err := HomeDir()
+	if err != nil {
+		return
+	}
 
 	// Parse the kube config path
 	kubeConfigPath = flag.String("kubeconfig", filepath.Join(home, ".kube", "config"), "(optional) absolute path to the kubeconfig file")
@@ -37,14 +38,10 @@ var _ = BeforeSuite(func() {
 
 	// use the current context in kubeconfig
 	config, err := clientcmd.BuildConfigFromFlags("", *kubeConfigPath)
-	Expect(err).To(BeNil())
+	if err != nil {
+		return
+	}
 
 	// create the clientset
-	kubeConfig, err = kubernetes.NewForConfig(config)
-	Expect(err).To(BeNil())
-})
-
-func TestSource(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Bootstrap")
+	return kube.NewForConfig(config)
 }
