+++ integration-tests/kubernetes/kubernetes_suite_test.go
@@ -1,0 +1,13 @@
+package kubernetes
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestSource(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Kubernetes")
+}
